0.2.9 - Added a Clone Scene option
- Allows used to decide if they want to export a clone of the scene, or
actually manupilate the original scene

+++ hifi_json_writer.py
@@ -28,7 +28,7 @@ import re
 import os
 import json
 
-from bpy.props import StringProperty
+from bpy.props import (StringProperty, BoolProperty)
 from mathutils import Quaternion
 from math import sqrt
 
@@ -56,7 +56,10 @@ def select(blender_object):
     
 
 def parse_object(blender_object, path, options):  
-    
+    # Store existing rotation mode, just in case.
+    stored_rotation_mode = bpy.context.object.rotation_mode
+    json_data = None
+    # Make sure context is quaternion for the models
     name = re.sub(r'\.\d{3}$', '', blender_object.name)
     # If you ahve an object thats the same mesh, but different object: All Objects will use this as reference allowing for instancing.
     mesh_name = blender_object.data.name
@@ -68,23 +71,25 @@ def parse_object(blender_object, path, options):
     orientation = quat_swap_nzy(blender_object.rotation_quaternion)
     position = swap_nzy(blender_object.location)
     
-    
     bpy.ops.object.select_all(action = 'DESELECT')
     if type == 'MESH':        
         blender_object.select = True
+        bpy.context.object.rotation_mode = 'QUATERNION'
         
         dimensions = swap_yz(blender_object.dimensions)
         
         bpy.ops.object.origin_set(type='GEOMETRY_ORIGIN', center='BOUNDS')
-        bpy.context.object.rotation_mode = 'QUATERNION'
 
+        temp_rotation = blender_object.rotation_quaternion
         # Temporary Rotate Model to a zero rotation so that the exported model rotation is normalized.
         blender_object.rotation_quaternion = Quaternion((1,0,0,0))
 
+        # TODO: Option to also export via gltf instead of fbx
         bpy.ops.export_scene.fbx(filepath=path + mesh_name + ".fbx", version='BIN7400', embed_textures=True, path_mode='COPY',
                                 use_selection=True, axis_forward='-Z', axis_up='Y')
 
-        blender_object.rotation_quaternion = blender_object.rotation_quaternion                    
+        # Restore earlier rotation
+        blender_object.rotation_quaternion = temp_rotation                    
         json_data = {
             'name': name,
             'id': scene_id,
@@ -133,9 +138,7 @@ def parse_object(blender_object, path, options):
             }
             
             json_data["parentID"] = str(parent_uuid)
-            
-            
-        return json_data
+        
             
     elif type == 'LAMP':
         print(name, 'is Light')
@@ -179,20 +182,21 @@ def parse_object(blender_object, path, options):
         }   
             
         if light.type is 'POINT':
-            blender_object.select = True
-            return json_data        
+            blender_object.select = True 
         
-        # TODO: Spot Lights require rotation by 90 degrees to get pointing in the right direction
-        return None
-        
+        # TODO: Spot Lights require rotation by 90 degrees to get pointing in the right direction        
     elif type == 'ARMATURE': # Same as Mesh actually.
         print(name, 'is armature')
     
     else:
         print('Skipping unsupported feature', name, type)
     
-        
+    
+    # Restore object's rotation mode
+    bpy.context.object.rotation_mode = stored_rotation_mode
+    
     bpy.ops.object.select_all(action = 'DESELECT')
+    return json_data
 
 # Rotation is based on the rotaiton of the parent and self. 
 def relative_rotation(parent_object):
@@ -222,14 +226,18 @@ class HifiJsonWriter(bpy.types.Operator, ExportHelper):
     directory = StringProperty()
     filename_ext = ".hifi.json"
     filter_glob = StringProperty(default="*.hifi.json", options={'HIDDEN'})
-    
+
     url_override = StringProperty(default="", name="Marketplace / Base Url", 
                                     description="Set Marketplace / URL Path here to override")
+    clone_scene = BoolProperty(default=True, name="Clone Scene prior to export", description="Clones the scene and performs the automated export functions on the clone instead of the original")
+    
     
     def draw(self, context):
         layout = self.layout
         layout.label("Url Override: Add Marketplace / URL to make sure that the content can be reached.")
         layout.prop(self, "url_override")
+        layout.label("Clone scene: Performs automated actions on a cloned scene instead of the original.")
+        layout.prop(self, "clone_scene")
         
 
     def execute(self, context):
@@ -240,12 +248,14 @@ class HifiJsonWriter(bpy.types.Operator, ExportHelper):
             raise Exception("Set the Marketplace / URL to make sure that the content can be reached.")
         
         current_scene = bpy.context.scene
+        read_scene = current_scene
         # Creating a temp copy to do the changes in.
-        test = bpy.ops.scene.new(type='FULL_COPY')
-
-        new_scene = bpy.context.scene
-        new_scene.name = 'Hifi_Export_Scene'
-
+        if not self.clone_scene:
+            bpy.ops.scene.new(type='FULL_COPY')
+            read_scene = bpy.context.scene # sets the new scene as the new scene
+            read_scene.name = 'Hifi_Export_Scene'
+        
+        
         # Clone Scene. Then select scene. After done delete scene
         path = os.path.dirname(os.path.realpath(self.filepath)) + '/'
         
@@ -262,15 +272,18 @@ class HifiJsonWriter(bpy.types.Operator, ExportHelper):
         
         entities = []
         
-        for blender_object in new_scene.objects:
+        for blender_object in read_scene.objects:
             
             parsed = parse_object(blender_object, path, {'url': url})
             
             if parsed:
                 entities.append(parsed)        
-            
 
-        bpy.ops.scene.delete()
+        # Delete Cloned scene
+        #     
+        if not self.clone_scene:
+            bpy.ops.scene.delete()
+        
         hifi_scene = {
             'Version': EXPORT_VERSION,
             'Entities': entities
